+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [26.06rc2] - 2026-05-21: "Quantum-Resistant Lightning Channel"
+## [26.06] - 2026-06-01: "Quantum-Resistant Lightning Channel"
 
 This release was named by @enaples.
 
@@ -69,7 +69,7 @@ Note: You should always set `allow-deprecated-apis=false` to test for changes.
  - Protocol: use BOLT4's paranoid advice about doing constant-time error decryption. ([#9044])
  - Fix for `splicein`, `spliceout`, and `dev-splice` commands where channel balances included partial sats. ([#9097])
  - build: fix build errors with GCC 15 (Arch Linux). ([#9075])
-
+ - Plugins: xpay no longer crashes on circular bolt11 routehints. ([#9174])
 
 ### EXPERIMENTAL
 
@@ -89,6 +89,7 @@ Note: You should always set `allow-deprecated-apis=false` to test for changes.
 [#9078]: https://github.com/ElementsProject/lightning/pull/9078
 [#8981]: https://github.com/ElementsProject/lightning/pull/8981
 [#8751]: https://github.com/ElementsProject/lightning/pull/8751
+[#9174]: https://github.com/ElementsProject/lightning/pull/9174
 [#8807]: https://github.com/ElementsProject/lightning/pull/8807
 [#9037]: https://github.com/ElementsProject/lightning/pull/9037
 [#9119]: https://github.com/ElementsProject/lightning/pull/9119
@@ -100,7 +101,7 @@ Note: You should always set `allow-deprecated-apis=false` to test for changes.
 [#8942]: https://github.com/ElementsProject/lightning/pull/8942
 [#8922]: https://github.com/ElementsProject/lightning/pull/8922
 [#9136]: https://github.com/ElementsProject/lightning/pull/9136
-[26.06rc2]: https://github.com/ElementsProject/lightning/releases/tag/v26.06rc2
+[26.06]: https://github.com/ElementsProject/lightning/releases/tag/v26.06
 
 ## [26.04.1] - 2026-04-25: "Negative Routing Fees II"